move modules to `src/` directory
To be able to use the `api` without importing issues, this would be good to organize the project as it should be.
