+++ test/test_integration.py
@@ -11,9 +11,9 @@ import pytest
 
 
 def run():
-    from model.mives import Mives
-    from model.poem_builder import PoemBuilder
-    from model.filter import Filter
+    from src.model.mives import Mives
+    from src.model.poem_builder import PoemBuilder
+    from src.model.filter import Filter
     import random
 
     padrao_ritmico: str = "ABAB ABAB CDC CDC"
